do you even lift state up, bro?
so...

- move the state property up the
widget tree from where we'd like it to
be accessible downstream
- if the upstream widget is stateless,
convert that to stateful (dart will
complain about this when the variable
is moved into it, which helps)
- pass the state property, 'List tasks' in
this case, as a parameter to the
downstream widget that needs to use
it.
- use the property in downstream
widget to compute as usual.
- consider turning downstream widget
to stateless if it doesn't need to hold
state (which might be the case here)